fix: off by ten error sweet. A few javadoc fixes as well.

+++ boogie-arinc/src/main/java/org/mitre/tdp/boogie/arinc/model/ArincControlledAirspaceLeg.java
@@ -120,11 +120,11 @@ public final class ArincControlledAirspaceLeg implements ArincModel {
   /**
    * See {@link ArcDistance}
    */
-  private final Integer arcDistance;
+  private final Double arcDistance;
   /**
    * See {@link ArcBearing}
    */
-  private final Integer arcBearing;
+  private final Double arcBearing;
   /**
    * See {@link Rnp}
    */
@@ -314,11 +314,11 @@ public final class ArincControlledAirspaceLeg implements ArincModel {
     return Optional.ofNullable(arcOriginLongitude);
   }
 
-  public Optional<Integer> arcDistance() {
+  public Optional<Double> arcDistance() {
     return Optional.ofNullable(arcDistance);
   }
 
-  public Optional<Integer> arcBearing() {
+  public Optional<Double> arcBearing() {
     return Optional.ofNullable(arcBearing);
   }
 
@@ -426,8 +426,8 @@ public final class ArincControlledAirspaceLeg implements ArincModel {
     private Double longitude;
     private Double arcOriginLatitude;
     private Double arcOriginLongitude;
-    private Integer arcDistance;
-    private Integer arcBearing;
+    private Double arcDistance;
+    private Double arcBearing;
     private Double rnp;
     private Double lowerLimit;
     private String lowerUnitIndicator;
@@ -542,12 +542,12 @@ public final class ArincControlledAirspaceLeg implements ArincModel {
       return this;
     }
 
-    public Builder arcDistance(Integer arcDistance) {
+    public Builder arcDistance(Double arcDistance) {
       this.arcDistance = arcDistance;
       return this;
     }
 
-    public Builder arcBearing(Integer arcBearing) {
+    public Builder arcBearing(Double arcBearing) {
       this.arcBearing = arcBearing;
       return this;
     }
