+++ boogie-arinc/src/main/java/org/mitre/tdp/boogie/arinc/model/ArincRestrictiveAirspaceLeg.java
@@ -102,11 +102,11 @@ public final class ArincRestrictiveAirspaceLeg implements ArincModel {
   /**
    * See {@link ArcDistance}
    */
-  private final Integer arcDistance;
+  private final Double arcDistance;
   /**
    * See {@link ArcBearing}
    */
-  private final Integer arcBearing;
+  private final Double arcBearing;
   /**
    * See {@link Limit}
    */
@@ -284,11 +284,11 @@ public final class ArincRestrictiveAirspaceLeg implements ArincModel {
     return Optional.ofNullable(arcOriginLongitude);
   }
 
-  public Optional<Integer> arcDistance() {
+  public Optional<Double> arcDistance() {
     return Optional.ofNullable(arcDistance);
   }
 
-  public Optional<Integer> arcBearing() {
+  public Optional<Double> arcBearing() {
     return Optional.ofNullable(arcBearing);
   }
 
@@ -377,8 +377,8 @@ public final class ArincRestrictiveAirspaceLeg implements ArincModel {
     private Double longitude;
     private Double arcOriginLatitude;
     private Double arcOriginLongitude;
-    private Integer arcDistance;
-    private Integer arcBearing;
+    private Double arcDistance;
+    private Double arcBearing;
     private Double lowerLimit;
     private String lowerUnitIndicator;
     private Double upperLimit;
@@ -480,12 +480,12 @@ public final class ArincRestrictiveAirspaceLeg implements ArincModel {
       return this;
     }
 
-    public Builder arcDistance(Integer arcDistance) {
+    public Builder arcDistance(Double arcDistance) {
       this.arcDistance = arcDistance;
       return this;
     }
 
-    public Builder arcBearing(Integer arcBearing) {
+    public Builder arcBearing(Double arcBearing) {
       this.arcBearing = arcBearing;
       return this;
     }
